requirements.txt
Flask
gunicorn
google-generativeai==0.2.0
python-dotenv
requests